requirements.txt
discord.py==2.4.*
asyncpg==0.30.*
python-dotenv==1.1.*
audioop-lts==0.2.*
